moved messaging client extensions closer to related interface
